bug in setter

+++ src/part_2_readable_good/utilities/InventoryItem.java
@@ -1,4 +1,4 @@
 package part_2_readable_good.utilities;
 
-public record InventoryItem(int id, String product, String image, String description, int price, String currency) {
+public record InventoryItem(int id, String product, String image, String description, Double price, String currency) {
 }
